+++ bucket.go
@@ -45,14 +45,3 @@ func (n *node) GetBucket(tx *bolt.Tx, children ...string) *bolt.Bucket {
 
 	return b
 }
-
-// CreateBucketIfNotExists creates the bucket below the current node if it doesn't
-// already exist.
-func (s *DB) CreateBucketIfNotExists(tx *bolt.Tx, bucket string) (*bolt.Bucket, error) {
-	return s.root.CreateBucketIfNotExists(tx, bucket)
-}
-
-// GetBucket returns the given bucket below the current node.
-func (s *DB) GetBucket(tx *bolt.Tx, children ...string) *bolt.Bucket {
-	return s.root.GetBucket(tx, children...)
-}
